pyproject.toml
[tool.poetry]
name = "neural"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.poetry.dependencies]
python = "^3.6"
numpy = "^1.17"
pillow = "^6.2"
scikit-learn = "^0.21.3"
scipy = "^1.3"
poetry = "^0.12.17"
joblib = "^0.14.0"
opencv-python = "^4.1"
matplotlib = "^3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
